backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

fastapi==0.109.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Login==0.6.2
psycopg2-binary==2.9.5
pydantic==1.10.2
pytest==7.2.1
SQLAlchemy==1.4.45
Werkzeug==2.2.2
